+++ src/lib/components/filter/components/index.ts
@@ -1,0 +1,1 @@
+export { FilterToolbarComponent } from './filter-toolbar/filter-toolbar.component';