+++ src/lib/components/filter/filter.component.ts
@@ -2,6 +2,9 @@ import { Component, EventEmitter, Inject, Output } from '@angular/core';
 import { L10nLocale, L10N_LOCALE } from 'angular-l10n';
 
 /**
+ * @deprecated since v11.
+ * Use <terra-filter-toolbar> instead.
+ *
  * @author mkunze
  * @description This component provides the default template and functionality to display form fields which are supposed to set filters
  */