+++ src/lib/components/filter/models/index.ts
@@ -1,0 +1,1 @@
+export { DisplayWithFn } from './display-with-function.interface';