+++ src/lib/components/filter/pipes/index.ts
@@ -1,0 +1,1 @@
+export { ChipDisplayValuePipe } from './chip-display-value.pipe';